+++ src/seyond_lidar_ros/src/driver/driver_lidar.h
@@ -1,5 +1,5 @@
 /**********************************************************************************************************************
-Copyright (c) 2025 Seyond
+Copyright (c) 2023 Seyond
 All rights reserved
 
 By downloading, copying, installing or using the software you agree to this license. If you do not agree to this
@@ -37,6 +37,7 @@ USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #include <string>
 #include <vector>
 #include <thread>
+#include <queue>
 #include <mutex>
 #include <fstream>
 #include <condition_variable>
@@ -57,50 +58,52 @@ typedef pcl::PointXYZI SeyondPoint;
 namespace seyond {
 
 struct CommonConfig {
-  std::string log_level;
-  bool fusion_enable;
-  std::string fusion_topic;
+  std::string log_level = "info";
+  bool fusion_enable = false;
+  std::string fusion_topic = "/iv_points_fusion";
+  int32_t fusion_time_window = 50;
+  int32_t fusion_buffer_size = 10;
 };
 
 struct LidarConfig {
   int32_t index;
-  bool replay_rosbag;
-  bool packet_mode;
-  int32_t aggregate_num;
+  bool replay_rosbag = false;
+  bool packet_mode = false;
+  bool enable_imu_msg = false;
+  int32_t aggregate_num = 20;
 
-  std::string frame_id;
-  std::string packet_topic;
-  std::string frame_topic;
+  std::string frame_id = "seyond";
+  std::string packet_topic = "/iv_packet";
+  std::string frame_topic = "/iv_points";
 
-  std::string lidar_name;
-  std::string lidar_ip;
-  int32_t port;
-  int32_t udp_port;
-  bool reflectance_mode;
-  int32_t multiple_return;
-  bool enable_falcon_ring;
-  bool enable_imu_msg;
+  std::string lidar_name = "seyond";
+  std::string lidar_ip = "172.168.1.10";
+  int32_t port = 8010;
+  int32_t udp_port = 8010;
+  bool reflectance_mode = true;
+  int32_t multiple_return = 1;
+  bool enable_falcon_ring = false;
 
-  bool continue_live;
+  bool continue_live = false;
 
-  std::string pcap_file;
-  std::string hv_table_file;
-  int32_t packet_rate;
-  int32_t file_rewind;
+  std::string inno_pc_file = "";
+  std::string pcap_file = "";
+  std::string hv_table_file = "";
+  int32_t packet_rate = 10000;
+  int32_t file_rewind = 0;
 
-  double max_range;
-  double min_range;
-  std::string name_value_pairs;
-  int32_t coordinate_mode;
+  double max_range = 2000.0;  // unit: meter
+  double min_range = 0.1;     // unit: meter
+  int32_t coordinate_mode = 3;
 
-  bool transform_enable;
-  double x;
-  double y;
-  double z;
-  double pitch;
-  double yaw;
-  double roll;
-  std::string transform_matrix;
+  bool transform_enable = false;
+  double x = 0.0;
+  double y = 0.0;
+  double z = 0.0;
+  double pitch = 0.0;
+  double yaw = 0.0;
+  double roll = 0.0;
+  std::string transform_matrix = "";
 };
 
 struct TransformParam {
@@ -112,12 +115,17 @@ struct TransformParam {
   double roll;
 };
 
+struct FrameData {
+  pcl::PointCloud<SeyondPoint>::Ptr cloud_ptr;
+  double timestamp;
+};
+
 class DriverLidar {
  public:
   explicit DriverLidar(const LidarConfig& lidar_config);
   ~DriverLidar();
 
-  // static callback warpper
+  // static callback wrapper
   static void lidar_message_callback_s(int32_t handle, void *ctx, uint32_t from_remote, enum InnoMessageLevel level,
                                        enum InnoMessageCode code, const char *error_message);
   static int32_t lidar_data_callback_s(int32_t handle, void *ctx, const InnoDataPacket *pkt);
@@ -137,9 +145,11 @@ class DriverLidar {
       const std::function<void(pcl::PointCloud<SeyondPoint> &, double)> &callback) {
     frame_publish_cb_ = callback;
   }
+
   void register_publish_imu_callback(const std::function<void(std::vector<float> &, uint64_t)>& callback) {
     imu_data_publish_cb_ = callback;
   }
+
   void init_transform_matrix();
   void transform_pointcloud();
   void convert_and_parse(const int8_t *pkt);
@@ -157,11 +167,12 @@ class DriverLidar {
   bool setup_lidar();
   int32_t lidar_live_process();
   int32_t pcap_playback_process();
-  int32_t set_config_name_value();
+  int32_t inno_pc_playback_process();
   void start_check_datacallback_thread();
   void data_packet_parse(const InnoDataPacket *pkt);
   template <typename PointType>
   void point_xyz_data_parse(bool is_use_refl, uint32_t point_num, PointType point_ptr);
+  void start_publish_thread();
 
  public:
   LidarConfig param_;
@@ -170,6 +181,13 @@ class DriverLidar {
   std::vector<char> anglehv_table_;
 
   pcl::PointCloud<SeyondPoint>::Ptr pcl_pc_ptr;
+  uint64_t reserve_size_{100000};
+
+  std::mutex pcl_pc_mutex_;
+  std::condition_variable pcl_pc_cv_;
+  std::queue<FrameData> frame_data_queue_;
+  std::thread publish_thread_;
+  std::atomic_bool publish_thread_running_{false};
 
   std::function<void(const int8_t*, uint64_t, double, bool)> packet_publish_cb_;
   std::function<void(pcl::PointCloud<SeyondPoint>&, double)> frame_publish_cb_;
@@ -183,7 +201,6 @@ class DriverLidar {
   std::mutex running_mutex_;
   std::atomic_bool is_receive_data_{false};
   int32_t lidar_handle_{-1};
-  bool fatal_error_{false};
   int64_t current_frame_id_{-1};
   std::vector<uint8_t> data_buffer;
   double current_ts_start_;