+++ src/seyond_lidar_ros/src/driver/ros2_driver_adapter.hpp
@@ -1,5 +1,5 @@
 /**********************************************************************************************************************
-Copyright (c) 2025 Seyond
+Copyright (c) 2024 Seyond
 All rights reserved
 
 By downloading, copying, installing or using the software you agree to this license. If you do not agree to this
@@ -30,7 +30,6 @@ USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #pragma once
 
 #include <pcl_conversions/pcl_conversions.h>
-
 #include <rclcpp/rclcpp.hpp>
 #include <sensor_msgs/msg/point_cloud2.hpp>
 #include <sensor_msgs/msg/imu.hpp>
@@ -119,8 +118,8 @@ class ROSAdapter {
       int64_t ts_ns = msg->timestamp * 1000;
       ros_msg.header.stamp.sec = ts_ns / 1000000000;
       ros_msg.header.stamp.nanosec = ts_ns % 1000000000;
-      ros_msg.width = driver_ptr_->pcl_pc_ptr->width;
-      ros_msg.height = driver_ptr_->pcl_pc_ptr->height;
+      ros_msg.width = driver_ptr_->pcl_pc_ptr->points.size();
+      ros_msg.height = 1;
       inno_frame_pub_->publish(std::move(ros_msg));
       driver_ptr_->pcl_pc_ptr->clear();
     }
@@ -168,7 +167,7 @@ class ROSAdapter {
     inno_frame_pub_->publish(std::move(ros_msg));
   }
 
-  void publishImu(const std::vector<float> imu_data, uint64_t imu_ts_ns) {
+  void publishImu(const std::vector<float>& imu_data, uint64_t imu_ts_ns) {
     sensor_msgs::msg::Imu imu_msg;
     imu_msg.header.frame_id = lidar_config_.frame_id;
     imu_msg.header.stamp.sec = imu_ts_ns / 1000000000;
@@ -248,6 +247,10 @@ class ROSNode {
     // common
     node_ptr_->get_parameter_or<std::string>("log_level", common_config_.log_level, "info");
     common_config_.fusion_enable = false;
+    // unused parameters, but still set default values
+    common_config_.fusion_topic = "/iv_points_fusion";
+    common_config_.fusion_time_window = 50;
+    common_config_.fusion_buffer_size = 10;
 
     // Parse parameters for ros
     node_ptr_->get_parameter_or<bool>("replay_rosbag", lidar_config.replay_rosbag, false);
@@ -269,14 +272,14 @@ class ROSNode {
 
     node_ptr_->get_parameter_or<bool>("continue_live", lidar_config.continue_live, false);
 
+    node_ptr_->get_parameter_or<std::string>("inno_pc_file", lidar_config.inno_pc_file, "");
     node_ptr_->get_parameter_or<std::string>("pcap_file", lidar_config.pcap_file, "");
     node_ptr_->get_parameter_or<std::string>("hv_table_file", lidar_config.hv_table_file, "");
     node_ptr_->get_parameter_or<int32_t>("packet_rate", lidar_config.packet_rate, 10000);
     node_ptr_->get_parameter_or<int32_t>("file_rewind", lidar_config.file_rewind, 0);
 
     node_ptr_->get_parameter_or<double>("max_range", lidar_config.max_range, 2000.0);  // unit: meter
-    node_ptr_->get_parameter_or<double>("min_range", lidar_config.min_range, 0.4);     // unit: meter
-    node_ptr_->get_parameter_or<std::string>("name_value_pairs", lidar_config.name_value_pairs, "");
+    node_ptr_->get_parameter_or<double>("min_range", lidar_config.min_range, 0.1);     // unit: meter
     node_ptr_->get_parameter_or<int32_t>("coordinate_mode", lidar_config.coordinate_mode, 3);
 
     node_ptr_->get_parameter_or<bool>("transform_enable", lidar_config.transform_enable, false);